Some work on how to handle replicas.
When a provider is disabled, the node is deactivated and all blockchains that had that node as signers is updated.

New1: Whe a provider is disabled, not only the signers list is updated but also the replicas list. Test for both Chromia0 and Enterprise0 is added for this.

New2: When a node becomes a signer, it is removed from the replicas list. So that a node cannot be both signer and replica at the same time.

Adding/removing replicas in Enterprise0: Only the provider can add/remove replicas.

Bug fix in rell code for applying new provider state.